+++ measure/measure.h
@@ -3,10 +3,10 @@
 
 namespace measure
 {
-    float Pa_mmHG(float pressure);
-    float mmHG_Pa(float pressure);
-    float pressureToAltitude(float seaLevel, float atmospheric);
-    float seaLevelForAltitude(float altitude, float atmospheric);
+    double Pa_mmHG(double pressure);
+    double mmHG_Pa(double pressure);
+    double pressureToAltitude(double seaLevel, double atmospheric);
+    double seaLevelForAltitude(double altitude, double atmospheric);
 }
 
 #endif // MEASURE_H
